Functional packer box generation!

+++ scripts/update.sh
@@ -1,10 +1,13 @@
-#!/usr/bin/env bash
+#!/bin/bash -eux
 # Disable the release upgrader
-echo "Disabling the release upgrader..."
+echo "==> Disabling the release upgrader"
 sed -i.bak 's/^Prompt=.*$/Prompt=never/' /etc/update-manager/release-upgrades
-echo "Updating before installing......."
-sudo apt-get update -y -qq >/dev/null 2>&1
-echo "Performing dist-upgrade (all packages and kernel)..." >/dev/null 2>&1
-apt-get -y dist-upgrade --force-yes >/dev/null 2>&1
-reboot
-sleep 60
+echo "==> Updating list of repositories"
+# apt-get update does not actually perform updates, it just downloads and indexes the list of packages
+apt-get -y update
+if [[ $UPDATE  =~ true || $UPDATE =~ 1 || $UPDATE =~ yes ]]; then
+    echo "==> Performing dist-upgrade (all packages and kernel)"
+    apt-get -y dist-upgrade --force-yes
+    reboot
+    sleep 60
+fi
